+++ custom_components/powercalc/strategy/lut.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+from decimal import Decimal
 
 import gzip
 import logging
@@ -98,7 +99,7 @@ class LutStrategy(PowerCalculationStrategyInterface):
         self._lut_registry = lut_registry
         self._model = model
 
-    async def calculate(self, entity_state: State) -> Optional[float]:
+    async def calculate(self, entity_state: State) -> Optional[Decimal]:
         """Calculate the power consumption based on brightness, mired, hsl values."""
         attrs = entity_state.attributes
         color_mode = attrs.get(ATTR_COLOR_MODE)
@@ -163,7 +164,7 @@ class LutStrategy(PowerCalculationStrategyInterface):
                 brightness,
             )
 
-        power = self.lookup_power(lookup_table, light_setting)
+        power = Decimal(self.lookup_power(lookup_table, light_setting))
         _LOGGER.debug("%s: Calculated power:%s", entity_state.entity_id, power)
         return power
